Refactor splitboxes into Views/
